refactor: move UDP codec into its own folder
